Reorg utils, rename validate -> validate_structure
