+++ tests/test_agnostic/test_utils.py
@@ -7,34 +7,34 @@ import numpy as np
 import pytest
 import zarr
 
-from geff.utils import validate
+from geff import validate_structure
 
 if TYPE_CHECKING:
     from pathlib import Path
 
 
-def test_validate(tmp_path: Path) -> None:
+def test_validate_structure(tmp_path: Path) -> None:
     # Does not exist
     with pytest.raises(FileNotFoundError, match=r"Path does not exist: does-not-exist"):
-        validate("does-not-exist")
+        validate_structure("does-not-exist")
 
     # remote zarr path does not raise existence error
     remote_path = "https://blah.com/test.zarr"
     with pytest.raises(ValueError, match=r"store must be a zarr StoreLike"):
-        validate(remote_path)
+        validate_structure(remote_path)
 
     # Path exists but is not a zarr store
     non_zarr_path = tmp_path / "not-a-zarr"
     non_zarr_path.mkdir()
     with pytest.raises(ValueError, match=r"store must be a zarr StoreLike"):
-        validate(non_zarr_path)
+        validate_structure(non_zarr_path)
 
     zpath = tmp_path / "test.zarr"
     z = zarr.open_group(zpath)
 
     # Missing metadata
     with pytest.raises(ValueError, match="No geff key found in"):
-        validate(zpath)
+        validate_structure(zpath)
     z.attrs["geff"] = {
         "geff_version": "0.0.1",
         "directed": True,
@@ -44,26 +44,26 @@ def test_validate(tmp_path: Path) -> None:
 
     # No nodes
     with pytest.raises(ValueError, match="'graph' group must contain a group named 'nodes'"):
-        validate(zpath)
+        validate_structure(zpath)
     nodes = z.create_group("nodes")
 
     # Nodes missing ids
     with pytest.raises(ValueError, match="'nodes' group must contain an 'ids' array"):
-        validate(zpath)
+        validate_structure(zpath)
     n_node = 10
     z["nodes/ids"] = np.zeros(n_node)
 
     # Nodes must have a props group
     with pytest.raises(ValueError, match="'nodes' group must contain a group named 'props'"):
-        validate(zpath)
+        validate_structure(zpath)
     nodes.create_group("props")
 
     # Subgroups in props must have values
     nodes.create_group("props/score")
     with pytest.raises(ValueError, match="Node property group 'score' must have a 'values' array"):
-        validate(zpath)
+        validate_structure(zpath)
     z["nodes/props/score/values"] = np.zeros(n_node)
-    validate(zpath)
+    validate_structure(zpath)
 
     # Property shape mismatch
     z["nodes/props/badshape/values"] = np.zeros(n_node * 2)
@@ -74,7 +74,7 @@ def test_validate(tmp_path: Path) -> None:
             f"which does not match id length {n_node}"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
 
     del z["nodes/props"]["badshape"]
     # Property missing shape mismatch
@@ -87,7 +87,7 @@ def test_validate(tmp_path: Path) -> None:
             f"which does not match id length {n_node}"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     del z["nodes/props"]["badshape"]
 
     # No edge group is okay, if the graph has no edges
@@ -95,7 +95,7 @@ def test_validate(tmp_path: Path) -> None:
 
     # Missing edge ids
     with pytest.raises(ValueError, match="'edges' group must contain an 'ids' array"):
-        validate(zpath)
+        validate_structure(zpath)
 
     # ids array must have last dim size 2
     n_edges = 5
@@ -107,7 +107,7 @@ def test_validate(tmp_path: Path) -> None:
             f"edges ids must have a last dimension of size 2, received shape {badshape}"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     del z["edges"]["ids"]
     z["edges/ids"] = np.zeros((n_edges, 2))
 
@@ -120,7 +120,7 @@ def test_validate(tmp_path: Path) -> None:
             f"which does not match id length {n_edges}"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     del z["edges/props/badshape"]["values"]
 
     # Property missing shape mismatch
@@ -133,7 +133,7 @@ def test_validate(tmp_path: Path) -> None:
             f"which does not match id length {n_edges}"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     del z["edges/props/badshape"]["missing"]
 
     # Nodes: property metadata has no matching data
@@ -147,7 +147,7 @@ def test_validate(tmp_path: Path) -> None:
         ValueError,
         match="Node property prop1 described in metadata is not present in props arrays",
     ):
-        validate(zpath)
+        validate_structure(zpath)
 
     # Nodes: inconsistent property metadata dtype
     z["nodes/props/prop1/values"] = np.zeros(n_node, dtype=np.float32)
@@ -159,7 +159,7 @@ def test_validate(tmp_path: Path) -> None:
             "metadata dtype <class 'numpy.int64'>"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     # Another type of dtype mismatch
     z["nodes/props/prop2/values"] = np.zeros(n_node, dtype="int16")
     with pytest.raises(
@@ -169,7 +169,7 @@ def test_validate(tmp_path: Path) -> None:
             "metadata dtype <class 'numpy.int64'>"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     z["nodes/props/prop2/values"] = np.zeros(n_node, dtype="int")  # clean up
 
     # Edges: property metadata has no matching data
@@ -181,7 +181,7 @@ def test_validate(tmp_path: Path) -> None:
         ValueError,
         match="Edge property prop3 described in metadata is not present in props arrays",
     ):
-        validate(zpath)
+        validate_structure(zpath)
 
     # Edges: inconsistent property metadata dtype
     z["edges/props/prop3/values"] = np.zeros(n_edges, dtype=np.float32)
@@ -192,7 +192,7 @@ def test_validate(tmp_path: Path) -> None:
             r"metadata dtype <class 'numpy.bool_*'>"
         ),
     ):
-        validate(zpath)
+        validate_structure(zpath)
     z["edges/props/prop3/values"] = np.zeros(n_edges, dtype="bool")  # clean up
 
     # No error raised when property with no matching prop metadata
@@ -205,4 +205,4 @@ def test_validate(tmp_path: Path) -> None:
     z.attrs["geff"] = geff_attrs
 
     # Everything passes
-    validate(zpath)
+    validate_structure(zpath)
